Fix thread safety issue in SSLEngine.
Use a 6 byte array instead of a 3 byte arrays.
The first 3 bytes are used by encrypt and the second
three bytes are used by decrypt.

We could write each of the bytes individually and
avoid the need for an array but that's 3 JNI calls instead
of 1.

TESTED : libcore.javax.net.*

bug: 11334293
Change-Id: I8c5d4b37e3f5d65938a6647e60ac69defe77e386