Offer PKCS#7 padding for AES and DES.
This offers PKCS#7 padding for all Cipher transformations which
currently support PKCS#5 padding.

PKCS#5 padding is a special case of PKCS#7 padding. PKCS#5 padding
is defined specifically for 64 bit long blocks. However, lots of code
assumes that PKCS#5 for other block sizes works exactly like PKCS#7,
and thus uses PKCS#5 padding where PKCS#7 should actually be used
(e.g., with AES). Thus, the same padding implementation will be
offered regardless of whether PKCS#5 or PKCS#7 padding is requested.

As an added benefit, this change speeds up by an order of magnitude
AES encryption and decryption when PKCS#7 padding is requested on
Android. This is because prior to this change AES with PKCS#7 padding
was by default backed by BouncyCastle, and with this change it is
backed by the much faster OpenSSL implementation.

Change-Id: I0ca8a952c67bc7aff172e22bd730378d41438067