Update patch set 2
Patch Set 2:

Yes, this is the behavior of the RI. Also, the contract of SSLContext.init does not say that SSLContext.init should blow up if it can't find an X509TrustManager or X509KeyManager in the provided parameters. Thus, the current implementation in Android is wrong.

Patch-set: 2
Label: Verified=0