Update patch set 1
Patch Set 1:

The motivation is that JSSE packages are soon going to need a level of awareness as to which user they run under. Most immediate use-case is each profile on a device trusting a different set of CAs.

But OK, moving out pieces of a self-contained library is a bad idea.

Patch-set: 1
Reviewer: Gerrit User 1045544 <1045544@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0