Enable PSK cipher suites when PSKKeyManager is provided.
This enables TLS-PSK cipher suites iff SSLContext is initialized with
a PSKKeyManager. For consistency, X.509 based cipher suites are no
longer enabled by default at all times -- they are now only enabled
by default iff SSLContext is initialized with a X509KeyManager or a
X509TrustManager.

When both X.509 and PSK cipher suites need to be enabled, PSK cipher
suites are given higher priority in the resulting list of cipher
suites. This is based on the assumption that in most cases users of
TLS/SSL who enable TLS-PSK would prefer TLS-PSK to be used when the
peer supports TLS-PSK.

Bug: 15073623
Change-Id: I8e2bc3e7a1ea8a986e468973b6bad19dc6b7bc3c