Update patch set 1
Patch Set 1:

I'm not sure that this is quite complete: for example I think I may need to delete OpenSSLEngine.java and references to that, but it's at least in a state where rough comments are applicable. (E.g. "we can't support supporting $x because $y needs it").

Patch-set: 1
Label: Verified=0