Update patch set 2
Patch Set 2: Code-Review+1

I think this is probably a good idea to be safe, but I'm not convinced that it's needed.

My thought process when removing this (and sorry for remembering when I commented on the other CL) was that I don't think it's needed. OpenSSL will seed itself from /dev/urandom and the other RAND_ calls are a waste.

That's certainly true of BoringSSL but, in hindsight, maybe it's not worth worrying about for OpenSSL and we should just leave this as-is.

(If you're following along, ssleay_rand_bytes in crypto/rand/md_rand.c calls RAND_poll if |initialized| is zero and RAND_poll, if you skip the VOS and OpenBSD sections, loads from |DEVRANDOM|, which is defined to include /dev/urandom in e_os.h.

Patch-set: 2
Reviewer: Gerrit User 1019036 <1019036@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1