Update patch set 6
Patch Set 6:

It's unclear to me how far we want to go with this approach to fixing premature finalization. 

Plusses:
It works now, and handles most or all cases we know about.

Minusses:
It adds complexity to the code and slows it down (appreciably due to added localrefs?) where abstractly neither should be required.

The theoretician in me cringes, in that this is a JNI-based workaround, which leverages a (undocumented) JNI implementation weakness, to solve what's fundamentally a pure Java language issue.  Even the long-proposed, but never implemented keepAlive(ref) Java primitive seems much cheaper and arguably cleaner.

In cases like this, it probably makes sense to do this now.  More generally, we should probably discuss how to proceed.

Patch-set: 6
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0