Enable any opaque private keys to be used with TLS/SSL stack.
Prior to this CL, opaque private keys (those that do not expose/export
their key material) were not fully supported by SSLSocket,
SSLServerSocket and SSLEngine implementations offered by Conscrypt.
In particular, opaque private keys which are backed by providers other
than Conscrypt were not supported.

This CL fixes this issue. Conscrypt's TLS/SSL stack now works with
arbitrary opaque private keys provided that:
* for EC private key: an instance of NONEwithECDSA Signature accepts
  the key for signing.
* for RSA private key: an instance of NONEwithRSA Signature accepts
  the key for signing and an instance of RSA/ECB/PKCS1Padding Cipher
  accept the key for decryption.
This normally requires that the JCA Provider which produced the
PrivateKey instance expose the above Cipher transform and Signature
algorithms.

HOW THIS WORKS

The underlying OpenSSL TLS/SSL stack uses the provided private keys
only to decrypt and sign. For opaque private keys these requests are
delegated (same as before, via CryptoUpcalls) to corresponding Cipher
(RSA/ECB/PKCS1Padding) and Signature (NONEwithRSA or NONEwithECDSA)
implementations.

In addition to the PrivateKey instances, OpenSSL needs the modulus and
order, supposedly to estimate output size of signing or decryption
operations. This information is not available via the PrivateKey
interface. However, an opaque private key may still implement the
RSAKey or ECKey interface which provides access to modulus or order
but does not provide access to key material. Moreover, in all use
cases of private keys with Conscrypt's TLS/SSL stack the modulus or
order can be obtained and provided to OpenSSL. In the case of private
keys used for client or server authentication, the public key of the
certificate is used as the source of the information. In the case of
TLS Channel ID, the order is currently fixed and known (only NIST
P-256 is supported).

Bug: 19284418
Change-Id: Iae716311637fa4275a768def92b8cff7e91a20cb