Enable opaque private keys to be used with TLS/SSL stack.
Prior to this CL, opaque private keys (instance of PrivateKey but
not instances of ECPrivateKey or RSAPrivateKey) were not supported by
SSLSocket, SSLServerSocket and SSLEngine implementations offered by
Conscrypt. This prevents the use of private keys which are backed by
providers other than Conscrypt and which do not expose/export their
key material (e.g., TEE-backed private keys).

This CL fixes this issue.

The fix is as follows. The underlying OpenSSL TLS/SSL stack uses
the provided private keys only to encrypt and sign. For opaque private
keys these requests are delegated (same as before, via CryptoUpcalls)
to corresponding Cipher and Signature implementations.

One complication is that OpenSSL still needs to know the modulus (for
RSA private keys) and order (for EC private keys), supposedly to
estimate output size. This information is not available via the
PrivateKey interface. Luckily, in all use cases of private keys with
Conscrypt's TLS/SSL stack this information can be obtained and
provided to OpenSSL. In the case of private keys used for client or
server authentication, the public key of the certificate is used as
the source of the information. In the case of TLS Channel ID, the
order is currently fixed and known (only NIST P-256 is supported).

Change-Id: Iae716311637fa4275a768def92b8cff7e91a20cb