Enable any opaque private keys to be used with TLS/SSL stack.
Prior to this CL, opaque private keys (those that do not expose/export
their key material) were not fully supported by SSLSocket,
SSLServerSocket and SSLEngine implementations offered by Conscrypt.
In particular, opaque private keys which are backed by providers other
than Conscrypt were not supported.

This CL fixes this issue. Conscrypt's TLS/SSL stack now accepts
arbitrary opaque private keys as long as modulus (for RSA keys) and
order (for EC keys) is known.

The underlying OpenSSL TLS/SSL stack uses the provided private keys
only to encrypt and sign. For opaque private keys these requests are
delegated (same as before, via CryptoUpcalls) to corresponding Cipher
(RSA/ECB/PKCS1Padding) and Signature (NONEwithRSA or NONEwithECDSA)
implementations.

Modulus and order are needed by OpenSSL, supposedly to estimate output
size of signing or encryption operations. This information is not
available via the PrivateKey interface. However, an opaque private key
may still implement the RSAKey or ECKey interface which provides
access to modulus or order but does not provide access to key
material. Moreover, in all use cases of private keys with Conscrypt's
TLS/SSL stack the modulus or order can be obtained and provided to
OpenSSL. In the case of private keys used for client or server
authentication, the public key of the certificate is used as the
source of the information. In the case of TLS Channel ID, the order is
currently fixed and known (only NIST P-256 is supported).

Bug: 19284418
Change-Id: Iae716311637fa4275a768def92b8cff7e91a20cb