Adjust Signature interface for BoringSSL
BoringSSL will push the BAD_SIGNATURE error onto the stack for every
signature error. In Java it just returns false from the Signature#verify
call when the signature is incorrect.  However, we still want to throw
an exception for raw RSA when the number of signature bytes is larger
than the modulus can express.

Bug: 21209646
Change-Id: Ie29eba8b12bb8d3220b7b29146c45b8dd416e338