Update patch set 2
Patch Set 2:

Maybe hold on a sec. Renegotiation is only mostly removed in BoringSSL. The case of a server requesting a client-certificate is still handled, although it may be disabled by default. If you're testing that situation then it may just be that conscrypt needs a function call added to enable it.

Patch-set: 2
Label: Verified=0