Add ability to prevent applications from overriding TrustManager
Modify conscrypt to force use of the default TrustManager when the
application's allowTLSTrustManagerOverride attribute is false.

As described by numerous sources including [1] [2] [3] and [4],
custom TrustManagers are a common source of security vulnerabilities.

This change goes along with corresponding changes to frameworks/base
and libcore with the same Change-Id.

[1] Tendulkar and Enck. An Application Package Configuration
Approach to Mitigating Android SSL Vulnerabilities. MoST 2014.
[2] Fahl, et al. Why Eve and Mallory Love Android: An Analysis
of Android SSL (In)Security. CCS '12.
[3] Sounthiraraj, et al. Large Scale, Automated Detection of SSL/TLS
Man-in-the-Middle Vulnerabilities in Android Apps. NDSS '14.
[4] FireEye. SSL Vulnerabilities: Who listens when Android
applications talk?
https://www.fireeye.com/blog/threat-research/2014/08/ssl-vulnerabilities-who-listens-when-android-applications-talk.html

Change-Id: If6ab0ab309488926cdde7fd55539bd059eea964d