Update patch set 5
Patch Set 5:

High level comment: you're doing this to support the SCT extension in a handshake? But the SCT extension is implemented natively in BoringSSL (and must be because it needs to store things in the SSL_SESSION) and thus trying to register a custom extension for the same thing will fail.

Patch-set: 5