Update patch set 5
Patch Set 5:

> High level comment: you're doing this to support the SCT extension
 > in a handshake? But the SCT extension is implemented natively in
 > BoringSSL (and must be because it needs to store things in the
 > SSL_SESSION) and thus trying to register a custom extension for the
 > same thing will fail.

PS 1 of the SCT implentation in conscrypt[1] used boringssl's implementation, but you mentioned a while ago ([2] and in some emails) that this API was going to be removed in favour of the more generic one.
And yes, I have to disable boringssl's SCT support for it to work.

1: https://android-review.googlesource.com/#/c/161222/1
2: https://android-review.googlesource.com/#/c/161224/

Patch-set: 5
Reviewer: Gerrit User 1077764 <1077764@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0