Update patch set 1
Patch Set 1:

> Either way sounds fine, I suppose.

I think it's better just to drop it, and so have sent https://android-review.googlesource.com/#/c/170788/ to you for review.

Your point about the timeout test makes sense, I think, but I still don't think that |wbuf.left| needs to be checked. The documentation for SSL_read[1] notes what happens when an error is returned and it's ok that the write buffer contains data in that case.

[1] https://boringssl.googlesource.com/boringssl/+/e8814df6998c8b44e77292926d975cee9dd9f743/include/openssl/ssl.h#326

Patch-set: 1