Update patch set 2
Patch Set 2:

The reason for the failure was that EVP_MD_CTX_copy fails if the digest struct inside EVP_MD_CTX is not initialized. The fix is to invoke EVP_MD_CTX_copy inside clone() only when mDigestInitializedInContext is true.

I re-uploaded this change (with the necessary adjustments) as https://android-review.googlesource.com/#/c/179992/. It's passing all the digest and signature tests.

Patch-set: 2