Fix Signature.sign broken on x86_64.
Conscrypt's Signature impl's sign() method invoked EVP_DigestSignFinal
with an uninitialized length of the output buffer. I misread the
method's documentation and assumed that the output buffer length does
not need to be initialized because it's an output-only parameter.
However, the parameter is an in-out parameter. This change fixes the
issue by initializing the parameter to the correct value.

Change-Id: Id5e205f185b9edf7189c26d0dd2f5a7c84e7c1c1