Update patch set 1
Patch Set 1:

I think what Kenny is getting at is that you it's not a requirement for hashCode to produce the same output as RI. What is a requirement is that all implementations of java.security.cert.Certificate inside Android produce the same output for the same encoded form, while ART is running. Once ART is restarted, the output of hashCode is permitted to change, but it still has to be consistent across all instances of Certificate.

Patch-set: 1