Update patch set 1
Patch Set 1:

If objects are equals(), they should have the same hashcode(). so if X509Certificate defines equals() so you can compare certs, the hashcode should be consistent as well.

generally speaking hashcodes being the same over different runs is not out of the ordinary, in fact for some classes the algorithm is specified in the documentation.

Patch-set: 1