Update patch set 2
Patch Set 2:

Thanks for the review. signapk.jar is using this module and your description sounds like the right approach is letting signapk.jar depend on libjavacore? I created

https://android-review.googlesource.com/#/c/214230/

before this patch, and I wasn't very sure which patch is better. My bet was this is the better choice, but it sounds I was wrong. That said, please note that libconscrypt_openjdk_jni.so won't be linked against libjavacore.so even with this patch as this change only adds LOCAL_REQUIRED_MODULES, not LOCAL_SHARED_LIBRALIES. I.e.,:

 $ ldd out/host/linux-x86/lib64/libconscrypt_openjdk_jni.so
        linux-vdso.so.1 =>  (0x00007fff1ff25000)
        libpthread.so.0 => /lib/x86_64-linux-gnu/libpthread.so.0 (0x00007fd044189000)
        libm.so.6 => /lib/x86_64-linux-gnu/libm.so.6 (0x00007fd043e83000)
        libgcc_s.so.1 => /lib/x86_64-linux-gnu/libgcc_s.so.1 (0x00007fd043c6d000)
        libc.so.6 => /lib/x86_64-linux-gnu/libc.so.6 (0x00007fd0438a8000)
        /lib64/ld-linux-x86-64.so.2 (0x00007fd0443cd000)

In other words, I think libconscrypt_openjdk_jni.so keeps working without libjavacore.so, but it won't build without javascript.

Patch-set: 2