Update patch set 2
Patch Set 2:

I'm wondering if missing libjavacore.so is safe. If not, I'd expect findAsynchronousCloseMonitorFuncs fails when dlopen fails. Anyway, I'll wait for Kenny's comment.

BTW, a correction for a silly typo in my previous comment: "but it won't build without javascript" => "but it won't build without libjavacore.so"

Patch-set: 2
Label: Verified=0