Substitute NULL for nullptr
Since we don't actually rely on a STL, we don't have <cstdlib> to get
NULL from, but since we're compiling C++11 we get nullptr for free.

This also fixes builds against MacOS SDK since it doesn't have <cstdlib>
available when you explicitly opt out of an STL in the Android.mk module.

Test: mmma -j32 external/conscrypt; make -j29 PRODUCT-sdk_phone_armv7-sdk
Change-Id: I54929c7e5c05ec271925f5f3d1896df1661e9b59