Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> What's the set of curves advertised by clients in elliptic_curves extension after this change?

The BoringSSL default which, for Android, is X25519, P-256, P-384, and P-521. Chrome uses the default as well, so you're in good company, although everything at Google except for Android takes P-521 out. We should do the same for you guys too, but there's no need to block this change on it.

(Keeping P-521 in Android was a temporary measure due to the problematic release process. Elsewhere we took it out to relieve complexity budget for X25519 which is a much much nicer curve.)

Patch-set: 1