Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thanks! How hard is it to add a CTS regression test for the contents of elliptic_curves extension in ClientHello? We should probably have one going forward.

I mean, you could hook up a memory BIO to the socket and parse out the ClientHello.

Although that kind of test is liable to break when we add new curves since we can't update the CTS repository and BoringSSL. We already had a previous issue here. Until Android gets its infrastructure up to speed, I think it would not be acceptable to add more of those tests. Rather you all should try to use BoringSSL's defaults which are maintained by those of us working on SSL across all of Google.

Remember that BoringSSL is not a static target, and Android is already in a very precarious situation w.r.t. BoringSSL changes because of just how difficult it is to update. I usually don't have the time to do it more than once every couple of months.

Patch-set: 1