Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Three reasons for CTS testing this:
> 1. the way Conscrypt uses Boring may have interesting unexpected effects on important aspects of TLS configuration, such as the issue this commit is fixing. We want to be in control over what the TLS configuration is. It doesn't mean we won't be changing it to remain modern and secure, it just means we don't want to discover these changes once a new build of Android has already shipped and users are reporting breakages.
> 2. On this note, CTS tests also constrain what OEMs can change in Android's TLS stack.
> 3. Android's platform-default TLS stack's use case is different from Chrome's. This leads to different defaults between the two. For example, the bar for dropping a weak cipher suite is different between Chrome and Android's TLS stack. Having said that, we ideally want to have the sets of defaults between Chrome and Android TLS stack pretty close whenever sensible.

Note BoringSSL's defaults also get used be everything else at Google. Chrome indeed can move first on a lot of things and we do that, by having Chrome temporarily configure things. Once a change has graduated through Chrome, we make the change globally. In some cases, we haven't need to do that so we don't even expose APIs to change the options.

Having multiple sets of configurations is actively harmful for our ability to manage SSL here, and historically Android has not been very good about keeping up with things. A CTS test is fine, although you should expect it to be broken regularly and it will be on you all to keep it up to date. But do not give us yet another set of configuration knobs that we need to watch and manage. We can manage that much better via BornigSSL.

Patch-set: 1