Remove NPN support
NPN is deprecated and clients should use ALPN instead for now. In the
interest of removing support for it in BoringSSL, we will remove NPN
support from Conscrypt as well.

For now keep around the NPN setters and getters in OpenSSLSocketImpl and
OpenSSLEngineImpl to help with backward compatibility for users that may
be using reflection.

Test: make -j32 && make -j32 cts && cts-tradefed run cts -d -m CtsLibcoreTestCases
Change-Id: Ia4edb21412d9c4b2440291ae0a8a97d2217bf5b5