Convert EVP_AEAD_CTX to stack-only
The EVP_AEAD_CTX object is only valid for the period of time in which it
is being used, so there is no need to have separate calls to
EVP_AEAD_CTX_init and EVP_AEAD_CTX_cleanup. Sweep the arguments to these
into the EVP_AEAD_CTX_seal and EVP_AEAD_CTX_open calls.

Test: make -j32 build-art-host vogar && vogar --mode host --classpath out/host/common/obj/JAVA_LIBRARIES/core-tests-support-hostdex_intermediates/classes.jack --classpath out/host/common/obj/JAVA_LIBRARIES/core-tests-hostdex_intermediates/classes.jack libcore/luni/src/test/java/libcore/javax/crypto/CipherTest.java
Change-Id: I02b2c5759efdd8f29ec354ee9c10053dd1f9c53c