Update patch set 7
Patch Set 7:

Sounds like this change is only needed as a workaround for Google Play Services Security Provider. We should not enabled this incorrect behavior in the platform's TLS stack. The contract of SSLSocket.setEnabledProtocols is that it throws IllegalArgumentException "when one or more of the protocols named by the parameter is not supported".

Patch-set: 7