OpenSSLEngineImpl: allow client renegotiations
The OpenSSLSocketImpl currently allows clients to initiate
renegotiations, but the OpenSSLEngineImpl had a bug that made this fail.
Allow client renegotiations and fix the bug.

The problem was that during a renegotiation the SSL_read_bio call would
be the one that output bytes related to the handshake. That was called
with localToRemoteSink, but then the handshaking part was expecting only
to see handshakeSink. This caused the first part of the renegotiation to
be lost because the localToRemoteSink was never sent over the wire to
the server.

The handkshakeSink appeared to be used to determine when a handshake was
completed, but this is not necessary because we receive a callback when
BoringSSL declares that the handshake is completed.

Test: cts-tradefed run cts -m CtsLibcoreTestCases -a arm64-v8a
Bug: 31773450
Change-Id: Id9bef5d1d71f28e9e666867855fa8f6cac0f4057