Update patch set 1
Patch Set 1:

> > Do we need to do this? Can't we just set -source 6 -target 6 ?
 > 
 > Er, sorry, just -target 6

Not sure that will address the problem.  The reason for this change is that this code caused errors in Android Studio's linter when attempting to set the min SDK version to 9.  There are other problems as well, but they are in the unbundled android code and we can address them separately.

Patch-set: 1
Reviewer: Gerrit User 1125908 <1125908@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0