OpenSSLEngineImpl: clear the handshakeSession near setting it
handshakeSession is only valid when calling out to certificate
verifiers. Certificate verification happens during the call to
SSL_do_handshake via the cert_verify_callback in
org_conscrypt_NativeCrypto.cpp which creates a JNI call into the
verifyCertificateChain Java language method.

During the JNI call to verifyCertificateChain, handshakeSession is set
before calling out to the X509TrustManager to verify the certificate
chain. To maintain symmetry in setting and unsetting of
handshakeSession, dereference it at the conclusion of
verifyCertificateChain.

Test: cts-tradefed run cts -m CtsLibcoreOkHttpTestCases -a arm64-v8a
Test: cts-tradefed run cts -m CtsLibcoreTestCases -a arm64-v8a
Change-Id: I9aa35cf01d2c15e7a23d1e87f2417009902b9091