Update patch set 13
Patch Set 13:

(7 comments)

This looks mostly reasonable, with the caveat that I'm not familiar with the locking or state machine and can't usefully comment on it. Though given the weirdnesses we discussed about the following commit this is a step towards, do you think it'd make more sense to go with the "all SSLSessions are SSLs" plan[*] or is this still a useful stepping stone towards it? (Also just to make sure: are you on board with that plan?)

[*] Concretely, I'll propose we do:
- Only have one SSLSession implementation and make it basically this CL's OpenSSLTransientSession
- Return that throughout the lifetime of the connection
- invalidate() is implemented by going to the session cache and blindly dumping everything associated with that cache key.
- SSL_SESSIONs going in and out of the session cache is managed internally by code completely detached from the public API.

Patch-set: 13
Reviewer: Gerrit User 1031048 <1031048@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0