Update patch set 1
Patch Set 1:

> Not sure how to review this. Is there a particular test in
 > CtsLibcoreTestCases that checks that all of these work? Or can you
 > suggest any other verification steps to convince myself that all of
 > these mappings are correct / that the work process you followed
 > will necessarily lead to a correct outcome?

There's libcore.java.security.ProviderTest that tests a number of things, including that all aliases point to something sensible, and there's the test I've added to it in the other change I sent that tests that Conscrypt covers all the BC aliases (and is where I got this list of aliases that didn't).  The OIDs are kind of annoying, but a Google search on them will show what they point to so you can verify that what they're being aliased to is correct.

Patch-set: 1
Reviewer: Gerrit User 1042885 <1042885@85c56323-6fa9-3386-8a01-6480fb634889>