Update patch set 1
Patch Set 1:

Thanks for the patch.  This package (including the makefile) is actually maintained on GitHub, so please instead submit this as a pull request to https://github.com/google/conscrypt and it'll make its way into AOSP when we next do an upstream sync.

Patch-set: 1
Reviewer: Gerrit User 1042885 <1042885@85c56323-6fa9-3386-8a01-6480fb634889>