Prefer Libcore.os over Os outside of android.* packages.
Libcore contains a mix usages of (a) android.system.Os, and
(b) libcore.io.Libcore.os.

This CL changes usages of (a) into usages of (b). This is a pure
refactoring, with no behavior change other than the call stack
becoming slightly shorter because the static methods in (a)
delegate directly to the instance methods in (b).

This reduces the dependency of non-android.* code onto android.*
code in libcore, and also makes libcore code more consistent.

Test: Treehugger
Bug: 67901714
Change-Id: I707ed2ed1e18bb829e693ea0836d0ccffed955df