Update patch set 1
Patch Set 1:

Sorry for the delay, I/O has been busy.

It looks like one of the tests that we recently added to Conscrypt uses a class that's not available on Android, so it's failing to build.  I can get that fixed, but it's probably easier to get you moving if you just cherry-pick the one change instead of doing a full upstream merge.

Patch-set: 1
Reviewer: Gerrit User 1042885 <1042885@85c56323-6fa9-3386-8a01-6480fb634889>