Use generated source not jarjar
This commit switches to generating source code for
com.android.org.conscrypt used on Android devices rather than relying
on jarjar to repackage the bytecode.

Source code can (and must) be regenerated using scripts in
external/conscrypt/srcgen when changes are made to bouncycastle
source code.

Having source code for the Android internal conscrypt is being done
to enable the code to be included in stubs for system and other internal
API surfaces: the stubs generation relies on having source (which can be
annotated, or just have associated API metadata).

Test: build
Test: CtsLibcoreTestCases
Bug: 111055375
Change-Id: Ic4780a3afd55c25009e79e1b07a713bd4ae3f8f8