Conscrypt: Fix OpenJDK9 build.
The conscrypt build target includes code from the package
com.android.org.conscrypt, which is also present in
core-intra-stubs-system-modules' version of java.base.
This broke compilation of the EXPERIMENTAL_USE_OPENJDK9=true
(OpenJDK 9) continuous build.

As a quick fix, this CL adds:

   patch_module: "java.base",

which considers the conscrypt code as part of java.base during
compilation. Longer term, we may want to get rid of
core-intra-stubs-system-modules instead, but for now this at
least fixes compilation.

Bug: 115604102
Test: EXPERIMENTAL_USE_OPENJDK9=true make conscrypt

Change-Id: Ibf8c785c0ad5e3ff4af1cabb22f01a13a0c653ba