Update patch set 5
Patch Set 5:

(1 comment)

This change seems quite reasonable as this target is only used on the host so there is no point in building it for Android. Maybe the name should be conscrypt-unbundled instead of conscrypt-host so that if it becomes necessary to use this on Android for some reason (e.g. to ensure that it does not cause issues for apps that want to bundle their own version) there is no need to change existing code. I think that you will also need to update the libcore/CleanSpec.mk file as that mentions conscrypt-nojarjar.jar (twice for some reason).

Patch-set: 5