Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> > Sending this out a bit early (treehugger hasn't validated it for me) but I'd like to know if you can think of a good reason why these have a static dep on conscrypt...?
> 
> Nope, the only thing I can think of is at some point maybe someone ran them not on a device?  Otherwise, I think the static dep doesn't do anything, since the platform copy of Conscrypt will take precedence over a bundled copy, I believe.

Yes - you're right about precedence. It is the main reason I'm changing them because it's not correct. I assumed the same about usage: that somebody in history probably wanted to run on host but didn't want to include --classpath to point to the conscrypt jar. Or it was just a mistake.

Patch-set: 1
Reviewer: Gerrit User 1038039 <1038039@85c56323-6fa9-3386-8a01-6480fb634889>