Add missing default constructors
Some IntraCoreApi annotations were missing as they were targeted at
default constructors and so were not present in the source. This change
adds the default constructors which allows the missing annotations to
be added.

Note, this does not actually affect the intra core api definition file
as that already includes the default constructor for these classes.
Raised 119859108 as this behavior may be a bug.

Bug: 117818301
Test: make checkbuild
Change-Id: I3fac495a784efb68ff117a1d0a36f0b799cdc49b