Revert "Remove CompatibilityCloseMonitor (#613)"
Broke com.squareup.okhttp.CallTest#cancelDuringHttpsConnect
(but not com.squareup.okhttp.CallTest#cancelDuringHttpConnect).

My theory is that the change caused threads waiting on a secure
socket to not wake up when the socket was closed.
ServerSocketConcurrentCloseTest only uses plain (non-TLS) sockets
so would not have noticed this. It might be worth expanding it
to cover secure sockets in a follow-up CL.

This reverts commit ae12b7d3d1b69a95c975d8ef59d9d976b32dd753.

Bug: 123121395
Test: Checked that the following fails before but passes after this CL:
      atest CtsLibcoreOkHttpTestCases:com.squareup.okhttp.CallTest#cancelDuringHttpsConnect

Change-Id: I5e340cafaf43c3d8316ca88c596d1a1d5c3c1d74