Make Platform.getDefaultProviderName public on Android.
Now that Conscrypt is on the boot class path, the reflexive call[1]
from a CTS test to a package private method fails due to
differing class loaders.  This change solves it by making the
method public (but not visible to the SDK), or we could also
do it by just using setAccessible before calling if that
seems preferable.

[1] https://android.googlesource.com/platform/external/conscrypt/+/master/
repackaged/testing/src/main/java/com/android/org/conscrypt/TestUtils.java#190

Bug: 123182286
Test: atest CtsLibcoreTestCases:com.android.org.conscrypt.ct.CTVerifierTest
Test: atest CtsLibcoreTestCases:com.android.org.conscrypt.CertPinManagerTest
Change-Id: I2f3fac6b09d689cd1312af3fbc2504ec35a5a959