CTS Fix: relax test for cipher suite selected during handshake.
The intent of test_SSLEngine_getHandshakeSession_duringHandshake is
to ensure that the handshake SSLSession can be read and is valid,
not to test a specific cypher suite was chosen.  However an
unrelated bug, http://b/135739091 means that on platforms with no
hardware AES the expected cipher suites are returned in the wrong
order.

Needs backporting to Github.

Bug: 135739091
Fixes: 129974850
Test: atest CtsLibcoreTestCases:com.android.org.conscrypt.javax.net.ssl.SSLEngineTest
Change-Id: I2654a7864a01528132436416b7b964cb7e55d424