Update patch set 1
Patch Set 1:

This was meant to be for libandroidio headers, but currently Conscrypt is pulling that in via dlopen/dlsym so isn't using it.  I suspect the right thing to do here is actually have the platform Conscrypt compile directly against libandroidio (or its stubs?) if we can make the runtime linking work for that, and only use dlopen from the unbundled Conscrypt.

If that seems workable to you then file a bug against me to do the work...

Patch-set: 1
Reviewer: Gerrit User 1334152 <1334152@85c56323-6fa9-3386-8a01-6480fb634889>