Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This was meant to be for libandroidio headers, but currently Conscrypt is pulling that in via dlopen/dlsym so isn't using it.  I suspect the right thing to do here is actually have the platform Conscrypt compile directly against libandroidio (or its stubs?) if we can make the runtime linking work for that, and only use dlopen from the unbundled Conscrypt.
> 
> If that seems workable to you then file a bug against me to do the work...

I don't know enough about conscrypt to know whether that is the right thing to do. The issue isn't what you are depending on but rather how you are depending on it. The include_dirs property has been deprecated (in favour of local_include_dirs, export_include_dirs and cc_library_headers and associated properties) and is blocking progress on unbundling as it references a path that does not exist in the unbundled build and so needs to be removed.

As this does not break the build are you ok for this to be removed now? If so could you please +2 it.

Patch-set: 1
Reviewer: Gerrit User 1039338 <1039338@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0