Don't explicitly call BORINGSSL_self_test.
After this CL topic, BORINGSSL_self_test runs automatically when
libcrypto is loaded. It is no longer possible or necessary to
start it manually.

The self test will abort() if it fails; the only other failure case
that we need to deal with is if (due to some build misconfiguration)
the version of libcrypto that we're seeing isn't built for FIPS mode.

Bug: 137267623
Test: Treehugger

Change-Id: If4e9a515529fdb466ba19c790c44f1d28905908e