Don't explicitly call BORINGSSL_self_test.
After this CL topic, BORINGSSL_self_test runs automatically when
libcrypto is loaded. It is no longer possible or necessary to
start it manually.

The self test will abort() if it fails; the only other failure case
that we need to deal with is if (due to some build misconfiguration)
the version of libcrypto that we're seeing isn't built for FIPS mode.

To avoid running the self test more than once per
{ 32bit ABI, 64bit ABI } x { /system, /apex } copy of libcrypto,
this CL also places boringssl_self_test{32,64} into both
/system/bin and /apex/com.android.conscrypt/bin, to be invoked
by the logic in http://r.android.com/1108553

Bug: 137267623
Test: Treehugger

Change-Id: If4e9a515529fdb466ba19c790c44f1d28905908e