Update patch set 1
Patch Set 1:

Will apps need to opt out too? If so then we would need to provide a public API for this.

The names and documentation of the methods being added do not make it very clear what these methods do and when they should be used. Having said that I don't really know conscrypt details so it may be that confusion is due to my limited knowledge.

Patch-set: 1
Reviewer: Gerrit User 1039338 <1039338@85c56323-6fa9-3386-8a01-6480fb634889>