Update patch set 1
Patch Set 1: Code-Review+2

To we have any unit tests for the old socket implementation's ability to set a handshake timeout? Asking because I'd expect that unit test (or any apps that might rely on it) to fail.

Doesn't block this CL, but might turn out the be a problem when switching the default over to be engine socket.

Patch-set: 1
Reviewer: Gerrit User 1103203 <1103203@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2