Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> FYI I don't understand what the empirical change after this CL is. Does it change the behavior of the dynamic linker so that libandroid.o is found at runtime, or does it affect where on the device libandroid.o lives, or what?

Nothing changes for libjavacrypto or other libraries. But the build system can follow the dependency graph when bundling APEX packages. What I intended is that, as I commented in the commit message, "com.android.conscrypt" APEX includes or lists its dependent libraries. In this case, because libjavacrypto in the conscrypt APEX requires libandroidio.so, the build system can follow the dependency and decide whether to include libandriodio.so as well in the apex _r to declare it as _required_ in the "requireNativeLibs" field in its APEX manifest. Here, libandroidio.so is supposed to be in the ART apex and it would be exposed by the build system as "provideNativeLibs" field in its APEX manifest.

Now linkerconfig(go/linker-config-apex) knows that the ART apex provides libandroidio.so and the conscrypt apex requires it by reading their APEX manifests. Our plan is to generate ld.config.txt for APEX inter-dependency.

I hope this may help.

Patch-set: 1
Reviewer: Gerrit User 1415877 <1415877@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0