Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (3 comments)

Also, sorry - looking at PS1, I can see there were plenty of discussions around the change of class names. I honestly don't know the right answer and I'm not overly familiar with the crypto provider stuff: if intra-core is supposed to document the things that have to be "visible" to the rest of "core", then those constructors are correct to be there. However, if the ART module doesn't inherently "know" them (i.e. it only knows about them because the conscrypt module told it the names), then I see why it would be safe to change them.

An extra factor here is that if the runtime folks start locking things down to enforce API surfaces (i.e. to prevent code in the ART module from using things outside of the intra-core API), then libcore *will* break if those classes are not "visible". Then it starts looking like the "is this a compatible change" logic needs to be fixed or we need to find a way to not have those members in the intra-core surface.

It makes me think that the factoring might be wrong - i.e. that libcore code shouldn't be asking for the names of things and then instantiating them. However, I suspect that might forced upon us by the way the crypto provider discovery works. Is the (I assume) Class.forName() behavior mandatory or could there be a middle ground where there's a method in the conscrypt module somewhere "instantiateFoo(String fooName)" that effectively does the "forName" call, maybe validates, and returns the Foo instance. That way the seam moves inside conscrypt module.

Patch-set: 2