Pare down Conscrypt exported intra-core API.
This removes almost all of Conscrypt's contribution to the
intra core API.  The following list of classes previously
in current-api.txt have been audited to ensure:
* There are no direct references outside of Conscrypt in git master
* There are no direct references outside of Conscrypt in the AOSP
Android 10 branch
* They can be instantiated using OpenSSLProvider
* There is no known unsupported app usage
That is to say that we can be quite quite confident that in Android 10 and
Android R they are only instantiated via the various public getInstance()
APIs. And in turn that means that references to the implementation
classes are all internal to the Conscrypt module and configured via
static data in OpenSSLProvider.  Other parts of the system do not depend
on the actual aboud the names or structure of these classes, just that
the services provided by OpenSSLProvider are consistent.

DESEDESecretKeyFactory
DefaultSSLContextImpl + subclasses
ECParameters
GCMParameters
IvParameters + subclasses
KeyGeneratorImpl + subclasses
OAEPParameters
OpenSSLAeadCipher + subclasses
OpenSSLCipher
OpenSSLCipherChaCha20
OpenSSLCipherRSA + subclasses
OpenSSLContextImpl + subclasses
OpenSSLECDHKeyAgreement
OpenSSLECKeyFactory
OpenSSLECKeyPairGenerator
OpenSSLEvpCipher
OpenSSLEvpCipherAES + subclasses
OpenSSLEvpCipherARC4
OpenSSLEvpCipherDESEDE + subclasses
OpenSSLMac + subclasses
OpenSSLRSAKeyFactory
OpenSSLRSAKeyPairGenerator
OpenSSLSignature + subclasses
OpenSSLSignatureRawECDSA
OpenSSLSignatureRawRSA
OpenSSLX509CertificateFactory
PSSParameters

Of the remaining classes: OpenSSLMessageDigestJDK and OpenSSLRandom are
used by non-Conscrypt tests and so should remain stable if possible,
although perhaps they belong in a different API.

OpenSSLRandom also has unsupported usage by a small number of apps, but
again maybe this should be recorded as such rather than keeping them in
the intra core API.

OpenSSLProvider was not previously recorded as being in the intra core
API when in fact it is the only real contributor to that API (as one
might expect).  It is used by reflection from sun.security.jca.Providers
and java.security.Providers.

For information only,  Bouncy Castle also has a class called GCMParameters
which performs the same functions for BC but is otherwise unrelated and
exists in a different package.

Bug: 134485888
Test: make
Change-Id: Ib5d77afb92cacd9077c678a7927933adc7cd1085