Update patch set 11
Patch Set 11: Code-Review+2

SGTM

Just for the context / to document discussion I had with prb@:

IIRC, the original idea of the intra-core .txt was to document all symbols where there were cross-module dependencies at runtime. i.e. anything where code in one core library module (e.g. conscrypt) was dependent at runtime on something in a different core library module (e.g. at the time, the embryonic runtime module) or vice versa.

I believe we added many of the entries that are now being removed after I suggested that any reflection-based runtime dependencies should be in the file too in order to record the full set of runtime dependencies. We obviously missed OpenSSLProvider by mistake.

The observation that allows the removal of entries is that some of the reflection-based (and other) dependencies are only discovered at runtime due to information from OpenSSLProvider, a class from the conscrypt module. We expect nobody to instantiate the entries being removed by hardcoding them. Therefore, the class names are free to change and need not be on the list of "things we have to keep fixed".

What this does mean is that the .txt file / API surface could not be used as the basis for any future runtime symbol enforcement / check; i.e. some of the symbols that the ART module needs at runtime will not be in the .txt file. If any such enforcement were needed in future, it would have to make use of information shipped inside the conscrypt module to account for renames. Since none of this is currently implemented (and may never be) it seems reasonable to drop anything from the list that isn't hardcoded (either directly or via reflection) in other core library modules.

Patch-set: 11
Reviewer: Gerrit User 1038039 <1038039@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2