Update patch set 1
Patch Set 1: -Autosubmit -Verified

Hmm, this is consistently hitting a tool failure with the device unresponsive after 30 minutes.

I noticed in local testing that CipherTest unlocks the device a lot, is it possible this can't be run as a presubmit, or needs to be run in a different way?

Patch-set: 1
Reviewer: Gerrit User 1334152 <1334152@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0
Label: Autosubmit=0
Label: Presubmit-Ready=0