Update patch set 1
Patch Set 1: -Autosubmit

> The description says this contains EXPECTED_BLOCK_SIZE change, but it doesn't, not sure how accurate that list is, but all changes LGTM

Oops, good catch.  The change after it in the list *was* included and I was careless while editting.

TH failure seems to be one corner-case test on cf86 only, no indication it affects real devices and am unable to reproduce on acloud to debug, so I'm going to mark it as a known failure for now.

Patch-set: 1
Label: Autosubmit=0
Label: Presubmit-Ready=0