Fix CheckJNI error
NativeCrypto_SSL_CIPHER_get_kx_name called to_SSL_CIPHER with
throwIfNull as true. If an exception was thrown it would be present
when the java string for the key exchange name.

Code paths already cataer for when an empty string is
returned by to_SSL_CIPER with throwIfNull as false.

Bug: 194980932
Test: adb root ; adb shell stop ; \
  adb shell setprop dalvik.vm.checkjni 1; \
  adb shell start; atest CtsLibcoreTestCases
Change-Id: I81229249c7b5e48e428c8828ebd74de82de712db