Revert "Use q launch defaults for dcla for com.android.conscrypt."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/33ab87fc-ee04-40de-99ad-7ae665fb0216). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib8274ef3c9b89a3b7710146d078ab9c84145ba46