Revert "Add conscrypt updatable certificates."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4a44bc4c-8cf9-40c0-a662-3c58d8e4d314). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib7f51d1374056607ca22927f94f59b1784005bf6