Revert "Add conscrypt updatable certificates."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a388af5-2459-441f-acc8-658fa502e028). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I271416a436d9a77945d1637ae6f36e9f4d27cccd