Revert "Add conscrypt updatable certificates."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b9153a1-1364-446b-9f54-9eee1eb68233). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I85cd3a2e50ca4f8ff447108d8152b374933839af