Revert "Add conscrypt updatable certificates."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ea79c8d9-d166-4a76-9154-251cb1630115). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I353b62ac3676755d145a98eae33100f53de689f5