Revert "Don't trigger TLS handshake until first I/O operation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5ab62dc6-8fe6-48d9-bcf0-be172084ce26). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie815b3e82817f1273caa365d9eb2f6b7a6dd599d