Revert "Don't trigger TLS handshake until first I/O operation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ff762cc8-b3fc-4eaa-ad4f-39fe9d077f26). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I10c1ec0bd7ba57a621d56523818876dd746d97ec