Revert "Don't trigger TLS handshake until first I/O operation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/03ebc1cd-855f-4c23-af5d-eeebb711491b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I83c8b87871c1f58d4df2ca102e7af55a142a4437