Revert "Don't trigger TLS handshake until first I/O operation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/26eb9a34-55da-4d60-a09c-150962cbd15a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib7fe1bb671f1d20e0f276db1c0aa5d403b36f8eb