Revert "Don't trigger TLS handshake until first I/O operation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/94c5141c-d75a-4d28-a34a-1f9dfced6714). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I918e4ae81af3eeead857835f1980987c0fb81496