Revert "Don't trigger TLS handshake until first I/O operation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/74368337-087b-4e9c-b23a-d2a667729c68). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I425c08c1ad47478e3f89c1bbfb95ea0cfb3db9ef